requirements.txt
requests
firebase-admin
google-generativeai
beautifulsoup4
lxml